# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BROKI INMOBILIARIA API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PropertyCategory =
  (typeof PropertyCategory)[keyof typeof PropertyCategory];

export const PropertyCategory = {
  economico: "economico",
  medio: "medio",
  premium: "premium",
} as const;

export interface Property {
  id: number;
  name: string;
  location: string;
  price: number;
  area: number;
  bedrooms: number;
  bathrooms: number;
  description: string;
  images: string[];
  category: PropertyCategory;
  available: boolean;
  createdAt: string;
}

export type CreatePropertyCategory =
  (typeof CreatePropertyCategory)[keyof typeof CreatePropertyCategory];

export const CreatePropertyCategory = {
  economico: "economico",
  medio: "medio",
  premium: "premium",
} as const;

export interface CreateProperty {
  name: string;
  location: string;
  price: number;
  area: number;
  bedrooms: number;
  bathrooms: number;
  description: string;
  images: string[];
  category: CreatePropertyCategory;
  available: boolean;
}

export interface SiteContent {
  heroTitle: string;
  heroSubtitle: string;
  heroCta: string;
  heroCtaSecondary: string;
  contactPhone: string;
  contactEmail: string;
  contactWhatsapp: string;
}