# lib/api-zod/src/generated/types/createProperty.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BROKI INMOBILIARIA API
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePropertyCategory } from "./createPropertyCategory";

export interface CreateProperty {
  name: string;
  location: string;
  price: number;
  area: number;
  bedrooms: number;
  bathrooms: number;
  description: string;
  images: string[];
  category: CreatePropertyCategory;
  available: boolean;
}