# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BROKI INMOBILIARIA API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyCategory } from "./propertyCategory";

export interface Property {
  id: number;
  name: string;
  location: string;
  price: number;
  area: number;
  bedrooms: number;
  bathrooms: number;
  description: string;
  images: string[];
  category: PropertyCategory;
  available: boolean;
  createdAt: Date;
}