# lib/api-zod/src/generated/types/siteContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BROKI INMOBILIARIA API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteContent {
  heroTitle: string;
  heroSubtitle: string;
  heroCta: string;
  heroCtaSecondary: string;
  contactPhone: string;
  contactEmail: string;
  contactWhatsapp: string;
}